fix tests failing because of community ordering
some tests were failing because the (l/e)comms were not generated in the same order as the expected output. after checking the pcap, it seems like pmacct (without netgauze) was sometimes outputting the communities in an order different than the one in the bgp attributes.

with pmacct-gauze the attribute order is preserved so we had to regenerate the expected output using it for the tests to pass.
before doing that it was checked that tests passed when using the previous expected output while ignoring the (l/e)comms so those only fields were the issue.

changes:
- change inconsistent daemon config to encode comms and aspath as arrays for those cases (some other tests still are not using arrays)
- change the expected output to be arrays
- regenerated output with correct comms ordering
- add tool script "arrayify.py" that converts expected an output file from string encoding to array encoding